Move iterators into own module.
