+++ src/decompress.rs
@@ -1,5 +1,4 @@
-use crate::{decode, instruction_length, DecodingError, Instruction, Register};
-use byteorder::{ByteOrder, LittleEndian};
+use crate::DecodingError;
 
 type DecompressionResult = Result<u32, DecodingError>;
 
@@ -100,104 +99,5 @@ pub fn decompress_q2(i: u16) -> DecompressionResult {
         0b110 /* C.SWSP */ => Err(DecodingError::Unimplemented),
         0b111 /* C.SDSP */ => Err(DecodingError::Unimplemented),
         _ => unreachable!(),
-    }
-}
-
-/// An iterator for all PC values where an instruction begins.
-pub struct LocationIter<'a> {
-    memory_view: &'a [u8],
-    current_index: u64,
-    address: u64,
-}
-
-impl LocationIter<'_> {
-    pub fn new(memory_view: &[u8], address: u64) -> LocationIter<'_> {
-        LocationIter {
-            memory_view,
-            current_index: 0,
-            address,
-        }
-    }
-
-    fn current_word(&self) -> u16 {
-        let idx: usize = self.current_index as usize;
-        let begin = &self.memory_view[idx..idx + 2];
-        LittleEndian::read_u16(begin)
-    }
-}
-
-impl Iterator for LocationIter<'_> {
-    type Item = u64;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match instruction_length(self.current_word()) {
-            2 => {
-                self.current_index += 2;
-                Some(self.address + self.current_index - 2)
-            }
-            4 => {
-                self.current_index += 4;
-                Some(self.address + self.current_index - 4)
-            }
-            l => panic!("Unimplemented instruction length: {}", l),
-        }
-    }
-}
-
-/// An iterator for all instructions in the program.
-pub struct InstructionIter<'a> {
-    memory_view: &'a [u8],
-    current_index: u64,
-}
-
-impl InstructionIter<'_> {
-    pub fn new(memory_view: &[u8]) -> InstructionIter<'_> {
-        InstructionIter {
-            memory_view,
-            current_index: 0,
-        }
-    }
-
-    fn current_word(&self) -> u16 {
-        let idx: usize = self.current_index as usize;
-        let begin = &self.memory_view[idx..idx + 2];
-
-        LittleEndian::read_u16(begin)
-    }
-
-    fn fetch_word(&mut self) -> u16 {
-        let word = self.current_word();
-
-        self.current_index += 2;
-
-        word
-    }
-
-    fn fetch_dword(&mut self) -> u32 {
-        let idx: usize = self.current_index as usize;
-        let begin = &self.memory_view[idx..idx + 4];
-
-        self.current_index += 4;
-
-        LittleEndian::read_u32(begin)
-    }
-}
-
-impl Iterator for InstructionIter<'_> {
-    type Item = Instruction;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.current_index >= self.memory_view.len() as u64 {
-            return None;
-        }
-
-        Some(
-            decode(match instruction_length(self.current_word()) {
-                2 => self.fetch_word().into(),
-                4 => self.fetch_dword(),
-                l => panic!("Unimplemented instruction length: {}", l),
-            })
-            .expect("valid instruction"),
-        )
     }
 }
